uint8 fix

+++ zlocalTest/p2p_mini_test/only_fpga.py
@@ -76,7 +76,7 @@ mod = TestFPGA
 target = [tvm.cpu(0), tvm.fpga(0)]
 vm = compile(mod, target)
 n=260
-inp_numpy = np.empty((n,), dtype="int32")
+inp_numpy = np.empty((n,), dtype="uint8")
 inp_numpy.fill(2)
 inp = tvm.nd.array(inp_numpy, target[0])
 res = vm["foo"](inp)
